feat(userconfigs): update user configs
