+++ charts/aiven-operator-crds/templates/aiven.io_serviceintegrations.yaml
@@ -111,6 +111,7 @@ spec:
                           - TSKV
                           - TSV
                           - TabSeparated
+                          - RawBLOB
                           type: string
                         group_name:
                           description: Kafka consumers group
